+++ packages/desktop/views/dashboard/collectibles/Collectibles.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { collectiblesRoute, CollectiblesRoute } from '@core/router'
-    import { NftDetailsView, CollectiblesGalleryView } from './views'
+    import { CollectiblesDetailsView, CollectiblesGalleryView } from './views'
 </script>
 
 <div class="w-full h-full flex flex-col flex-nowrap p-8 relative flex-1 bg-gray-50 dark:bg-gray-900">
@@ -8,6 +8,6 @@
         <CollectiblesGalleryView />
     {/if}
     {#if $collectiblesRoute === CollectiblesRoute.Details}
-        <NftDetailsView />
+        <CollectiblesDetailsView />
     {/if}
 </div>
