+++ packages/desktop/views/dashboard/collectibles/views/CollectiblesDetailsView.svelte
