+++ packages/desktop/views/dashboard/collectibles/views/index.js
@@ -1,2 +1,2 @@
-export { default as NftDetailsView } from './NftDetailsView.svelte'
+export { default as CollectiblesDetailsView } from './CollectiblesDetailsView.svelte'
 export { default as CollectiblesGalleryView } from './CollectiblesGalleryView.svelte'
